% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{coef_cutoff}
\alias{coef_cutoff}
\title{cut coefficients}
\usage{
coef_cutoff(coefs, beta_cutoff = 4, alpha_cutoff = 12)
}
\arguments{
\item{coefs}{list with coefficients}

\item{beta_cutoff}{numeric}

\item{alpha_cutoff}{numeric}
}
\value{
list of glm estimated beta, alpha coefficients with their p-values and adsusted q-values
}
\description{
cut coefficients
}
