% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{SeaBASS}
\alias{SeaBASS}
\title{Fit SeaBASS model}
\usage{
SeaBASS(
  data,
  initial = NULL,
  horseshoe = TRUE,
  cutoff_beta = 4,
  cutoff_alpha = 12,
  mu_alpha = NULL,
  mu_beta = NULL,
  a_alpha = 0.001,
  b_alpha = 0.001,
  init_inv_sigma2_beta = NULL,
  init_inv_sigma2_alpha = NULL,
  init_inv_a_beta = NULL,
  init_inv_a_tau = NULL,
  init_inv_tau2 = NULL,
  nu_0 = 1,
  eta2_0 = 1,
  nu_1 = 1,
  eta2_1 = 1,
  include_intercept = FALSE,
  ELBO_stop = 1,
  max_iter = 5000,
  ELBO_diff_tol = 1e-06,
  paras_diff_tol = 1e-06,
  verbose = 0,
  save_profile = 1
)
}
\arguments{
\item{data}{list with AE count (A), confounders (X), vaccine indication (V), number of each categories}

\item{initial}{list include the parameter initial values, default is from glm fit results}
}
\description{
Fit SeaBASS model
}
